+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This is intended as an overview of the major changes
 
 NEXT VERSION
 ===
-- ...
+- Support for the Gitlab Exclusion rules for the codeowners
 
 v.1.13.0
 ===